# mobile/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "城迹 CityPulse API",
    "description": "可信、结构化、可定位的城市动态事件 API。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/candidates": {
      "get": {
        "tags": [
          "admin-candidates"
        ],
        "summary": "List Candidates",
        "operationId": "list_candidates_api_v1_admin_candidates_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/CandidateReviewStatus",
              "default": "pending"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminCandidate"
                  },
                  "title": "Response List Candidates Api V1 Admin Candidates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/candidates/{candidate_id}/comparison": {
      "get": {
        "tags": [
          "admin-candidates"
        ],
        "summary": "Compare Candidate Source",
        "operationId": "compare_candidate_source_api_v1_admin_candidates__candidate_id__comparison_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateComparison"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/candidates/{candidate_id}/enrich": {
      "post": {
        "tags": [
          "admin-candidates"
        ],
        "summary": "Enrich Candidate",
        "operationId": "enrich_candidate_api_v1_admin_candidates__candidate_id__enrich_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Candidate Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateEnrich"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCandidate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/candidates/{candidate_id}/approve": {
      "post": {
        "tags": [
          "admin-candidates"
        ],
        "summary": "Approve Candidate",
        "operationId": "approve_candidate_api_v1_admin_candidates__candidate_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Candidate Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateApproval"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCandidate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/candidates/{candidate_id}/reject": {
      "post": {
        "tags": [
          "admin-candidates"
        ],
        "summary": "Reject Candidate",
        "operationId": "reject_candidate_api_v1_admin_candidates__candidate_id__reject_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Candidate Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateRejection"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCandidate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/corrections": {
      "get": {
        "tags": [
          "admin-corrections"
        ],
        "summary": "List Corrections",
        "operationId": "list_corrections_api_v1_admin_corrections_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/CorrectionStatus",
              "default": "pending"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminCorrection"
                  },
                  "title": "Response List Corrections Api V1 Admin Corrections Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/corrections": {
      "post": {
        "tags": [
          "corrections"
        ],
        "summary": "Create Correction",
        "operationId": "create_correction_api_v1_corrections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorrectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminCandidate": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "$ref": "#/components/schemas/EventCategory"
          },
          "organizer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organizer"
          },
          "price": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "venue_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "district": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "enrichment": {
            "$ref": "#/components/schemas/EnrichmentInfo"
          },
          "starts_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts At"
          },
          "ends_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends At"
          },
          "official_url": {
            "type": "string",
            "title": "Official Url"
          },
          "facts": {
            "additionalProperties": true,
            "type": "object",
            "title": "Facts"
          },
          "review_status": {
            "$ref": "#/components/schemas/CandidateReviewStatus"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "review_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Note"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "category",
          "organizer",
          "price",
          "venue_name",
          "address",
          "city",
          "district",
          "latitude",
          "longitude",
          "starts_at",
          "ends_at",
          "official_url",
          "facts",
          "review_status",
          "updated_at",
          "event_id",
          "reviewed_by",
          "reviewed_at",
          "review_note"
        ],
        "title": "AdminCandidate"
      },
      "AdminCorrection": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "evidence_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Url"
          },
          "contact_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "status": {
            "$ref": "#/components/schemas/CorrectionStatus"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "resolution_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Note"
          }
        },
        "type": "object",
        "required": [
          "id",
          "event_id",
          "kind",
          "message",
          "evidence_url",
          "contact_email",
          "status",
          "updated_at",
          "created_at",
          "reviewed_by",
          "resolution_note"
        ],
        "title": "AdminCorrection"
      },
...
      "CandidateApproval": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "category": {
            "$ref": "#/components/schemas/EventCategory"
          },
          "summary": {
            "type": "string",
            "maxLength": 360,
            "minLength": 1,
            "title": "Summary"
          },
          "description": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Description"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "title": "Starts At"
          },
          "ends_at": {
            "type": "string",
            "format": "date-time",
            "title": "Ends At"
          },
          "venue_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Venue Name"
          },
          "address": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Address"
          },
          "city": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "City"
          },
          "district": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "District"
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          },
          "organizer": {
            "type": "string",
            "maxLength": 180,
            "minLength": 1,
            "title": "Organizer"
          },
          "price": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "status": {
            "$ref": "#/components/schemas/EventStatus"
          },
          "evidence_url": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "format": "uri",
            "title": "Evidence Url"
          },
          "expected_updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expected Updated At"
          },
          "review_note": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Review Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "category",
          "summary",
          "description",
          "starts_at",
          "ends_at",
          "venue_name",
          "address",
          "city",
          "district",
          "latitude",
          "longitude",
          "organizer",
          "status",
          "evidence_url",
          "expected_updated_at",
          "review_note"
        ],
        "title": "CandidateApproval"
      },
      "CandidateComparison": {
        "properties": {
          "event_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "changes": {
            "items": {
              "$ref": "#/components/schemas/FieldChange"
            },
            "type": "array",
            "title": "Changes"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "changes"
        ],
        "title": "CandidateComparison"
      },
      "CandidateEnrich": {
        "properties": {
          "expected_updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expected Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_updated_at"
        ],
        "title": "CandidateEnrich"
      },
      "CandidateRejection": {
        "properties": {
          "expected_updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expected Updated At"
          },
          "review_note": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Review Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_updated_at",
          "review_note"
        ],
        "title": "CandidateRejection"
      },
      "CandidateReviewStatus": {
        "type": "string",
        "enum": [
          "pending",
          "approved",
          "rejected"
        ],
        "title": "CandidateReviewStatus"
      },
...
      "CorrectionCreate": {
        "properties": {
          "event_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "time",
              "location",
              "status",
              "source",
              "copyright",
              "other"
            ],
            "title": "Kind"
          },
          "message": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 10,
            "title": "Message"
          },
          "evidence_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Url"
          },
          "contact_email": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 320
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "message"
        ],
        "title": "CorrectionCreate"
      },
...
      "CorrectionResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CorrectionReceipt"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "CorrectionResponse"
      },
...
      "CorrectionStatus": {
        "type": "string",
        "enum": [
          "pending",
          "reviewing",
          "accepted",
          "rejected"
        ],
        "title": "CorrectionStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}